separate MariPowerTanker and DirectPowerBoat from Boat class
